requirements.txt
pyinstaller
python-socketio[asyncio]
python-socketio[asyncio_client]
eventlet
flask
requests
websocket-client
python-dotenv
pexpect
pytest